Create ConfigFileIO
