+++ Assets/CustomInputManager/Docs/2 InGame/Example.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8f64d5247310324428332319d064a6e9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
