+++ Assets/CustomInputManager/Scripts/ConfigFileIO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 41dea3dc84ba6c7468f7e9a55ea2ddf9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
