+++ Assets/CustomInputManager/Scripts/InputManager.cs
@@ -2,9 +2,9 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
+using UnityEditor;
 using UnityEngine;
 using UnityEngine.UI;
-using UnityEditor;
 
 ///Methods for Control changing menu, with integrated InputManager by Ben C.
 ///Since 2017 December
@@ -29,10 +29,13 @@ public class InputManager : MonoBehaviour {
 
 	[SerializeField]
 	///Full list of custom keybinds. Initialize with default controls.
-	private KeybindList controlList = new KeybindList();
+	private KeybindList controlList;
 
 	///Event for key detection
 	private Event currentEvent;
+
+	///Config file interface
+	private ConfigFileIO config;
 
 	/*
 	####################
@@ -44,32 +47,15 @@ public class InputManager : MonoBehaviour {
 	/// Awake is called when the script instance is being loaded.
 	/// </summary>
 	void Awake() {
+
+		//Init ConfigFileIO
+		config = new ConfigFileIO();
+		controlList = config.controlList;
+		defaultsPath = config.defaultsPath;
+		configPath = config.configPath;
+
 		infoTextContent = "Press " + cancelKeyCode + " to cancel key selection";
-		configPath = Application.dataPath + "/CustomInputManager/Config/controls.cfg";
-		defaultsPath = Application.dataPath + "/CustomInputManager/Config/defaultcontrols.cfg";
 
-		if (!Directory.Exists(Application.dataPath + "/CustomInputManager/Config")) {
-			Debug.Log("Creating Config folder...");
-			AssetDatabase.CreateFolder("Assets/CustomInputManager", "Config");
-		}
-
-		//validate controls file
-		try {
-			LoadControls(configPath);
-			Debug.Log("Successfully loaded controls file");
-		} catch {
-			Debug.Log("Controls file is nonexistent or corrupted. Generating new one...");
-			if (File.Exists(defaultsPath)) {
-				LoadControls(defaultsPath);
-				Debug.Log("Loaded controls from defaultcontrols.cfg");
-			} else {
-				Debug.Log("Default Controls are nonexistent or corrupted. Generating new one...");
-				controlList.addKeybind(0, "SampleKey", "Space");
-				WriteControls(defaultsPath);
-			}
-
-			WriteControls(configPath);
-		}
 	}
 
 	/// <summary>
@@ -95,23 +81,6 @@ public class InputManager : MonoBehaviour {
 		currentEvent = Event.current;
 	}
 
-	///Writes controls from ControlList to the config file.
-	public void WriteControls(string filePath) {
-		using(var writer = new StreamWriter(File.Create(filePath))) {
-			writer.WriteLine(controlList);
-		}
-	}
-
-	///Loads controls from the config file to ControlList.
-	public void LoadControls(string filePath) {
-		string[] keys = File.ReadAllLines(filePath);
-		controlList.reset();
-		foreach (string key in keys)
-			if (key.Length > 0)
-				controlList.addKeybind(key);
-
-		// Debug.Log(controlList);
-	}
 
 	///While key is being set: Waits for a valid input
 	public IEnumerator WaitForKey(int id, InputButton btnRef) {
@@ -145,7 +114,7 @@ public class InputManager : MonoBehaviour {
 					infoText.enabled = false;
 
 					//Save to file
-					WriteControls(configPath);
+					config.WriteControls(config.configPath);
 
 					yield break;
 				} else yield return null;
@@ -157,24 +126,23 @@ public class InputManager : MonoBehaviour {
 		controlList.generateButtons();
 	}
 
-    /*
+	/*
 	#######
 	# API #
 	#######	
 	 */
 
-    /// <summary>
-    /// Returns the Input Manager in the scene 
-    /// Assign to an InputManager object in your script.
-    /// </summary>
-    public static InputManager GetInputManager() {
-        try {
-            return (InputManager)FindObjectOfType(typeof(InputManager));
-        }
-        catch {
-            throw new NullReferenceException("InputManager could not be found!");
-        }
-    }
+	/// <summary>
+	/// Returns the Input Manager in the scene 
+	/// Assign to an InputManager object in your script.
+	/// </summary>
+	public static InputManager GetInputManager() {
+		try {
+			return (InputManager)FindObjectOfType(typeof(InputManager));
+		} catch {
+			throw new NullReferenceException("InputManager could not be found!");
+		}
+	}
 
 	///<summary> Returns true if the given key is pressed down</summary>
 	public bool GetKey(string name) {
@@ -191,20 +159,18 @@ public class InputManager : MonoBehaviour {
 		return Input.GetKeyUp(GetKeyCode(name).ToLower());
 	}
 
-    ///<summary> Returns the KeyCode corresponding to the keybind with given name. No case conversion.</summary>
-    public string GetKeyCode(string name) {
+	///<summary> Returns the KeyCode corresponding to the keybind with given name. No case conversion.</summary>
+	public string GetKeyCode(string name) {
 		return controlList.getKeybind(name).keyCode;
 	}
 
-    ///<summary> Returns the KeyCode corresponding to the keybind with given ID</summary>
-    public string GetKeyCode(int id) {
+	///<summary> Returns the KeyCode corresponding to the keybind with given ID</summary>
+	public string GetKeyCode(int id) {
 		return controlList.getKeybind(id).keyCode;
 	}
 
-    ///<summary> Wipes user preferences and copies default configuration to user configuration</summary>
-    public void ResetControls() {
-		LoadControls(defaultsPath);
-		WriteControls(configPath);
-		Debug.Log("Reset Controls");
+	///<summary> Wipes user preferences and copies default configuration to user configuration</summary>
+	public void ResetControls() {
+		config.ResetControls();
 	}
 }
