+++ ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,1 @@
-m_EditorVersion: 2018.1.0b7
+m_EditorVersion: 2018.1.0b6
